% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfb3_template.R
\name{gfb3_template}
\alias{gfb3_template}
\title{Return a GFB3 template tibble with correct column types}
\usage{
gfb3_template(n = 0L)
}
\arguments{
\item{n}{Integer. Number of empty rows to include (default 0).}
}
\value{
A tibble with all GFB3 columns and correct data types.
}
\description{
Return a GFB3 template tibble with correct column types
}
