% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/this_script.R
\name{this_script}
\alias{this_script}
\title{Get the path of the currently active script}
\usage{
this_script()
}
\value{
A character string with the full path to the active script.
}
\description{
Returns the file path of the script currently open and active in RStudio.
Intended for use with \code{\link{rm_from_script}} and
\code{\link{rm_except_script}}.
}
\note{
Only works inside RStudio. Will fail if called from a non-RStudio
  environment or when a script is run via \code{source()}.
}
\examples{
\dontrun{
this_script()
rm_from_script(this_script())
}

}
